requirements.txt
# Data & ML
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning
torch>=2.1.0
tensorflow>=2.15.0

# Reinforcement Learning
stable-baselines3>=2.1.0
gymnasium>=0.29.0  # Replaces deprecated gym

# Time Series
prophet>=1.1.4
statsmodels>=0.14.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.0
python-multipart>=0.0.6

# Simulation
simpy>=4.0.0
mesa>=2.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Additional ML
torch-geometric>=2.4.0

# Code Quality
black>=23.7.0
flake8>=6.1.0